Use modularity on EL8
Now that there's a proper module on EL8 this should be used.

In practice a user may see conflicts if the Ruby 2.5 or PostgreSQL
module has been enabled. Those are the default modules but foreman (and
thus katello) pulls in ruby:2.7 and postgresql:12. This means a user
must run that version.

In case of PostgreSQL it can be that there's already a version 10
database. In that case an upgrade must be performed, but this is not
documented. This can be considered unsupported for installations.

However, there are essentially no real upgrade instructions anyway. It's
all part of foreman-maintain but that may result in a chicken-and-egg
situation where you would need updated modules to update. So for now I'm
not touching that.

It also changes the troubleshooting guide appendix to be included in the
foreman-el installation. Then it links to that, which keeps the
quickstart guide shorter.