move preview deployment and commenting to a dedicated workflow
`pull_request` triggered workflows are lacking permissions to
- create comments
- access secrets

by preparing the data in the original workflow, but moving the actual
deployment to one that is triggered on `workflow_run`, we can avoid this
issue.

background: github (rightfully) limits the permissions as in a PR an
external attacker could craft a malicious update to the workflow and
gain more access than they should. `workflow_run` workflows run the code
from the default branch, not the PR, avoiding this attack vector.